+++ packages/utils/src/derived/number.ts
@@ -1,10 +1,14 @@
 import { SNI } from "../index";
 
 
-
 export function NumberDerived() {
-  // @ts-ignore
-  Number.prototype.Equal = function(this: number, val: string | number) {
+  Number.prototype.equal = function(this: number, val: string | number) {
     return SNI(this, val);
+  };
+
+  Number.prototype.log = function(this: number, mark?: string) {
+    const content = mark ? `${mark}: ${this}` : this;
+    console.log(content);
+    return content
   };
 }